adding exploratory parameter
Added an "exploratory" parameter to allow switching between user and LLM generated prompts. If set to True, the LLM will examine the user question and offer 4 possible methods for analysis. Will select the best method, and create a task list that will be sent to LLM as a prompt.
This works well for a vague user prompts but not so well for well defined prompts. The default is False