added answer ranking, prompts tweak
Added a functionality where the final answer is sent back to llm for ranking. This is a first step, to allow to build a KB in a vector db. Only  answers with a rank above threshold will be stored in the KB, and used in the subsequent iterations as a context.
Also transferred the promps to a new prompts.py module to make the main module cleaner.
The debug task now includes an instruction to check for any malicious or harming elements. This is to allow for post prompt check, to reduce the risk of prompt injection.